requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.3.0
chromadb==0.4.18
pypdf2==3.0.1
python-docx==1.1.0
markdown==3.5.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
colorama==0.4.6
requests==2.31.0
pdfkit==1.0.0
# 可选：用于搜索功能
# google-search-results==2.4.2
